Extract PlayerCards from Game.tsx
